requirements.txt
annotated-types==0.5.0
bcrypt==3.2.2
certifi==2022.5.18.1
colorama==0.4.6
cryptography==37.0.2
GeoAlchemy2==0.14.2
httpx==0.24.1
itsdangerous==2.1.2
kafka-python==2.0.2
psutil==5.9.1
psycopg2==2.9.3
pydantic==2.4.2
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
pydantic_core==2.10.1
Pygments==2.16.1
PyJWT==2.6.0
pyOpenSSL==22.0.0
python-dotenv==1.0.0
pytz==2022.1
requests==2.27.1
rich==13.6.0
httpx==0.24.1
SQLAlchemy==2.0.19
sqlparse==0.4.2
termcolor==1.1.0
terminado==0.15.0
typer==0.4.2
typing_extensions==4.8.0
tzdata==2022.1
urllib3==1.26.9
uvicorn==0.23.2
redis==5.0.1